virtual-study-group/requirements.txt
alembic==1.14.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Reuploaded==1.4.0
Flask-Reuploads==1.0.1.dev0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
watchdog==6.0.0
Werkzeug==3.1.3
WTForms==3.2.1
